# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/users": {
      "get": {
        "operationId": "AdminController_users",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminUserDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/users/{id}": {
      "patch": {
        "operationId": "AdminController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/audit": {
      "get": {
        "operationId": "AdminController_audit",
        "parameters": [
          {
            "name": "action",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditPageDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/admin/email-templates/{type}/{locale}": {
      "put": {
        "operationId": "AdminController_saveTemplate",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveTemplateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "operationId": "AdminController_resetTemplate",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/api/me/fines/{id}/pay-online": {
      "post": {
        "operationId": "PaymentsController_start",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnlineCheckoutDto"
                }
              }
            }
          }
        },
        "tags": [
          "payments"
        ]
      }
    },
    "/api/payments/webhook": {
      "post": {
        "operationId": "PaymentsController_webhook",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "payments"
        ]
      }
    },
    "/api/payments/{ref}": {
      "get": {
        "operationId": "PaymentsController_info",
        "parameters": [
          {
            "name": "ref",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentInfoDto"
                }
              }
            }
          }
        },
        "tags": [
          "payments"
        ]
      }
    },
    "/api/payments/mock/{ref}/complete": {
      "post": {
        "operationId": "PaymentsController_mockComplete",
        "parameters": [
          {
            "name": "ref",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "payments"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/verify-email": {
      "post": {
        "operationId": "AuthController_verify",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionUserDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgot",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "operationId": "AuthController_reset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/2fa/enable": {
      "post": {
        "operationId": "AuthController_enable2fa",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotpCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCodesDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/2fa/disable": {
      "post": {
        "operationId": "AuthController_disable2fa",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotpDisableDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "BiblioApp API",
    "description": "",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "AdminUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "MEMBER",
              "LIBRARIAN",
              "ADMIN"
            ]
          },
          "memberNumber": {
            "type": "string",
            "nullable": true
          },
          "disabled": {
            "type": "boolean"
          },
          "totpEnabled": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role",
          "memberNumber",
          "disabled",
          "totpEnabled",
          "createdAt"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "MEMBER",
              "LIBRARIAN",
              "ADMIN"
            ]
          },
          "disabled": {
            "type": "boolean"
          }
        }
      },
      "AuditItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "action": {
            "type": "string"
          },
          "userId": {
            "type": "number",
            "nullable": true
          },
          "userEmail": {
            "type": "string",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "action",
          "userId",
          "userEmail",
          "detail",
          "createdAt"
        ]
      },
      "AuditPageDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditItemDto"
            }
          }
        },
        "required": [
          "total",
          "page",
          "pageSize",
          "items"
        ]
      },
      "TemplateDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "locale": {
            "type": "string",
            "enum": [
              "nl",
              "en"
            ]
          },
          "subject": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "customized": {
            "type": "boolean"
          },
          "defaultSubject": {
            "type": "string"
          },
          "defaultBody": {
            "type": "string"
          },
          "placeholders": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "type",
          "locale",
          "subject",
          "body",
          "customized",
          "defaultSubject",
          "defaultBody",
          "placeholders"
        ]
      },
      "SaveTemplateDto": {
        "type": "object",
        "properties": {
          "subject": {
            "type": "string"
          },
          "body": {
            "type": "string"
          }
        },
        "required": [
          "subject",
          "body"
        ]
      },
      "OnlineCheckoutDto": {
        "type": "object",
        "properties": {
          "paymentId": {
            "type": "number"
          },
          "checkoutUrl": {
            "type": "string",
            "description": "Stuur de gebruiker hierheen om te betalen"
          }
        },
        "required": [
          "paymentId",
          "checkoutUrl"
        ]
      },
      "WebhookDto": {
        "type": "object",
        "properties": {
          "providerRef": {
            "type": "string"
          }
        },
        "required": [
          "providerRef"
        ]
      },
      "PaymentInfoDto": {
        "type": "object",
        "properties": {
          "providerRef": {
            "type": "string"
          },
          "amountCents": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PAID",
              "FAILED"
            ]
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "providerRef",
          "amountCents",
          "status",
          "description"
        ]
      },
      "CompleteDto": {
        "type": "object",
        "properties": {
          "outcome": {
            "type": "string",
            "enum": [
              "PAID",
              "FAILED"
            ]
          }
        },
        "required": [
          "outcome"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 10
          }
        },
        "required": [
          "email",
          "name",
          "password"
        ]
      },
      "TokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "required": [
          "token"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "totp": {
            "type": "string",
            "description": "6-cijferige 2FA-code of een herstelcode (alleen bij ingeschakelde 2FA)"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "SessionUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "MEMBER",
              "LIBRARIAN",
              "ADMIN"
            ]
          },
          "locale": {
            "type": "string",
            "enum": [
              "nl",
              "en"
            ]
          },
          "totpEnabled": {
            "type": "boolean"
          },
          "memberNumber": {
            "type": "string",
            "nullable": true
          },
          "csrfToken": {
            "type": "string",
            "description": "Stuur mee als X-CSRF-Token bij POST/PATCH/DELETE"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role",
          "locale",
          "totpEnabled",
          "memberNumber",
          "csrfToken"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 10
          }
        },
        "required": [
          "token",
          "password"
        ]
      },
...
      "TotpCodeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          }
        },
        "required": [
          "code"
        ]
      },
      "RecoveryCodesDto": {
        "type": "object",
        "properties": {
          "codes": {
            "description": "Eenmalig te tonen; bewaar ze veilig",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "codes"
        ]
      },
      "TotpDisableDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "password",
          "code"
        ]
      },
...
    }
  }
}